Crate tokio_c_mod

Infrastructure for creating tokio threads able to communicate with C event loop.

This provides a TokioCMod exposed to C with c_module! as tokio_c_module. A global [ArcSwapOption] stores the tokio runtime’s [Handle] allowing us to run a thread with a custom Future object.

Overview

The tokio_c_module is implemented with 3 methods:

initialize : initialize the module, with a shutdown_send signal and a runtime.

on_term : send the shutdown signal to the Tokio thread when the process receives a termination signal.

shutdown : terminate the thread and make sure it is finished.

The spawn function allows us to spawn a task using the tokio_c_module runtime and the Future input. It returns a [tokio::task::JoinHandle].

Example

Use tokio_c_module with main_c_queue_schedule

module_require(tokio_get_module());
let blocker = ElBlocker(unsafe { el_blocker_register() });

tokio_c_mod::spawn(async move {
    main_c_queue_schedule(move || {
        let mut blocker = blocker;
        unsafe {
            el_unregister(&raw mut blocker.0);
        }
    });
});

unsafe {
    el_loop();
}
module_release(tokio_get_module());

Functions

spawn

Spawn a future on the tokio_c_module runtime.

tokio_get_module